feat: add parse array func into utils.js and unit test
